+++ src/Model/Common/LicensorMeta.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Jikan\Model\Common;
-
-class LicensorMeta extends ItemBasicMeta
-{
-
-}