+++ src/Model/Common/StudioMeta.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Jikan\Model\Common;
-
-class StudioMeta extends ItemBasicMeta
-{
-
-}