+++ src/Model/Search/AnimeSearchListItem.php
@@ -2,7 +2,6 @@
 
 namespace Jikan\Model\Search;
 
-use Jikan\Model\Common\MalUrl;
 use Jikan\Parser;
 
 /**
@@ -101,16 +100,7 @@ class AnimeSearchListItem
         $instance->endDate = $parser->getEndDate();
         $instance->members = $parser->getMembers();
         $instance->rated = $parser->getRated();
-        $instance->airing =
-            null === $instance->endDate
-            && null !== $instance->startDate
-            &&
-            (
-                new \DateTimeImmutable(
-                    'now',
-                    new \DateTimeZone('UTC')
-                ) > $instance->startDate
-            );
+        $instance->airing = $parser->isAiring();
 
         return $instance;
     }