+++ src/MyAnimeList/MalClient.php
@@ -11,8 +11,8 @@
 
 namespace Jikan\MyAnimeList;
 
-use Goutte\Client;
 use GuzzleHttp\Client as GuzzleClient;
+use Jikan\Exception\BadResponseException;
 use Jikan\Exception\ParserException;
 use Jikan\Goutte\GoutteWrapper;
 use Jikan\Model;
@@ -47,8 +47,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeRequest $request
      * @return Model\Anime\Anime
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnime(Request\Anime\AnimeRequest $request): Model\Anime\Anime
     {
@@ -65,11 +65,22 @@ class MalClient
     /**
      * @param Request\Anime\AnimeEpisodesRequest $request
      * @return Model\Anime\Episodes
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeEpisodes(Request\Anime\AnimeEpisodesRequest $request): Model\Anime\Episodes
     {
+        // Episode page returns 404 when there are no results
+        try {
+            $crawler = $this->ghoutte->request('GET', $request->getPath());
+        } catch (\Exception $e) {
+            if ($e->getCode() === 404) {
+                return new Model\Anime\Episodes();
+            }
+
+            throw $e;
+        }
+
         $crawler = $this->ghoutte->request('GET', $request->getPath());
         try {
             $parser = new Parser\Anime\EpisodesParser($crawler);
@@ -83,8 +94,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeVideosRequest $request
      * @return Model\Anime\AnimeVideos
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeVideos(Request\Anime\AnimeVideosRequest $request): Model\Anime\AnimeVideos
     {
@@ -101,8 +112,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaRequest $request
      * @return Model\Manga\Manga
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getManga(Request\Manga\MangaRequest $request): Model\Manga\Manga
     {
@@ -116,15 +127,24 @@ class MalClient
         }
     }
 
+
     /**
      * @param Request\Character\CharacterRequest $request
      * @return Model\Character\Character
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getCharacter(Request\Character\CharacterRequest $request): Model\Character\Character
     {
         $crawler = $this->ghoutte->request('GET', $request->getPath());
+
+        // MAL returns `Invalid ID provided.` instead of 404 on invalid characters
+        $badResult = $crawler->filterXPath('//*[@id="content"]/div[@class="badresult"]');
+
+        if ($badResult->count()) {
+            throw new BadResponseException(sprintf('404 on %s', $request->getPath()), 404);
+        }
+
         try {
             $parser = new Parser\Character\CharacterParser($crawler);
 
@@ -137,8 +157,8 @@ class MalClient
     /**
      * @param Request\Person\PersonRequest $request
      * @return Model\Person\Person
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getPerson(Request\Person\PersonRequest $request): Model\Person\Person
     {
@@ -155,8 +175,8 @@ class MalClient
     /**
      * @param Request\User\UserProfileRequest $request
      * @return Model\User\Profile
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getUserProfile(Request\User\UserProfileRequest $request): Model\User\Profile
     {
@@ -173,8 +193,8 @@ class MalClient
     /**
      * @param Request\User\UserFriendsRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getUserFriends(Request\User\UserFriendsRequest $request): array
     {
@@ -191,8 +211,8 @@ class MalClient
     /**
      * @param Request\Seasonal\SeasonalRequest $request
      * @return Model\Seasonal\Seasonal
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getSeasonal(Request\Seasonal\SeasonalRequest $request): Model\Seasonal\Seasonal
     {
@@ -209,8 +229,8 @@ class MalClient
     /**
      * @param Request\Producer\ProducerRequest $request
      * @return Model\Producer\Producer
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getProducer(Request\Producer\ProducerRequest $request): Model\Producer\Producer
     {
@@ -227,8 +247,8 @@ class MalClient
     /**
      * @param Request\Magazine\MagazineRequest $request
      * @return Model\Magazine\Magazine
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMagazine(Request\Magazine\MagazineRequest $request): Model\Magazine\Magazine
     {
@@ -245,8 +265,8 @@ class MalClient
     /**
      * @param Request\Genre\AnimeGenreRequest $request
      * @return Model\Genre\AnimeGenre
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeGenre(Request\Genre\AnimeGenreRequest $request): Model\Genre\AnimeGenre
     {
@@ -263,8 +283,8 @@ class MalClient
     /**
      * @param Request\Genre\MangaGenreRequest $request
      * @return Model\Genre\MangaGenre
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaGenre(Request\Genre\MangaGenreRequest $request): Model\Genre\MangaGenre
     {
@@ -281,8 +301,8 @@ class MalClient
     /**
      * @param Request\Schedule\ScheduleRequest $request
      * @return Model\Schedule\Schedule
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getSchedule(Request\Schedule\ScheduleRequest $request): Model\Schedule\Schedule
     {
@@ -299,8 +319,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeCharactersAndStaffRequest $request
      * @return Model\Anime\AnimeCharactersAndStaff
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeCharactersAndStaff(
         Request\Anime\AnimeCharactersAndStaffRequest $request
@@ -318,8 +338,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimePicturesRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimePictures(Request\Anime\AnimePicturesRequest $request): array
     {
@@ -336,8 +356,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaPicturesRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaPictures(Request\Manga\MangaPicturesRequest $request): array
     {
@@ -354,8 +374,8 @@ class MalClient
     /**
      * @param Request\Character\CharacterPicturesRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getCharacterPictures(Request\Character\CharacterPicturesRequest $request): array
     {
@@ -373,6 +393,7 @@ class MalClient
      * @param Request\Person\PersonPicturesRequest $request
      *
      * @return Model\Common\Picture[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getPersonPictures(Request\Person\PersonPicturesRequest $request): array
@@ -391,6 +412,7 @@ class MalClient
      * @param Request\RequestInterface $request
      *
      * @return Model\News\NewsListItem[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getNewsList(Request\RequestInterface $request): array
@@ -409,6 +431,7 @@ class MalClient
      * @param Request\Search\AnimeSearchRequest $request
      *
      * @return Model\Search\AnimeSearch
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getAnimeSearch(Request\Search\AnimeSearchRequest $request): Model\Search\AnimeSearch
@@ -427,6 +450,7 @@ class MalClient
      * @param Request\Search\MangaSearchRequest $request
      *
      * @return Model\Search\MangaSearch
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getMangaSearch(Request\Search\MangaSearchRequest $request): Model\Search\MangaSearch
@@ -445,6 +469,7 @@ class MalClient
      * @param Request\Search\CharacterSearchRequest $request
      *
      * @return Model\Search\CharacterSearch
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getCharacterSearch(Request\Search\CharacterSearchRequest $request): Model\Search\CharacterSearch
@@ -467,7 +492,17 @@ class MalClient
      */
     public function getPersonSearch(Request\Search\PersonSearchRequest $request): Model\Search\PersonSearch
     {
-        $crawler = $this->ghoutte->request('GET', $request->getPath());
+        // Person page for some reason returns 404 when there are no results :facepalm:
+        try {
+            $crawler = $this->ghoutte->request('GET', $request->getPath());
+        } catch (\Exception $e) {
+            if ($e->getCode() === 404) {
+                return new Model\Search\PersonSearch();
+            }
+
+            throw $e;
+        }
+
         try {
             $parser = new Parser\Search\PersonSearchParser($crawler);
 
@@ -481,6 +516,7 @@ class MalClient
      * @param Request\Manga\MangaCharactersRequest $request
      *
      * @return Model\Manga\CharacterListItem[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getMangaCharacters(Request\Manga\MangaCharactersRequest $request): array
@@ -499,6 +535,7 @@ class MalClient
      * @param Request\Top\TopAnimeRequest $request
      *
      * @return Model\Top\TopAnime[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getTopAnime(Request\Top\TopAnimeRequest $request): array
@@ -517,6 +554,7 @@ class MalClient
      * @param Request\Top\TopMangaRequest $request
      *
      * @return Model\Top\TopManga[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getTopManga(Request\Top\TopMangaRequest $request): array
@@ -535,6 +573,7 @@ class MalClient
      * @param Request\Top\TopCharactersRequest $request
      *
      * @return Model\Top\TopCharacter[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getTopCharacters(Request\Top\TopCharactersRequest $request): array
@@ -553,6 +592,7 @@ class MalClient
      * @param Request\Top\TopPeopleRequest $request
      *
      * @return Model\Top\TopPerson[]
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getTopPeople(Request\Top\TopPeopleRequest $request): array
@@ -571,6 +611,7 @@ class MalClient
      * @param Request\Anime\AnimeStatsRequest $request
      *
      * @return Model\Anime\AnimeStats
+     * @throws BadResponseException
      * @throws ParserException
      */
     public function getAnimeStats(Request\Anime\AnimeStatsRequest $request): Model\Anime\AnimeStats
@@ -588,8 +629,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaStatsRequest $request
      * @return Model\Manga\MangaStats
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaStats(Request\Manga\MangaStatsRequest $request): Model\Manga\MangaStats
     {
@@ -606,8 +647,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeForumRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeForum(Request\Anime\AnimeForumRequest $request): array
     {
@@ -624,8 +665,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaForumRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaForum(Request\Manga\MangaForumRequest $request): array
     {
@@ -642,8 +683,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeMoreInfoRequest $request
      * @return string|null
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeMoreInfo(Request\Anime\AnimeMoreInfoRequest $request): ?string
     {
@@ -660,8 +701,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaMoreInfoRequest $request
      * @return string|null
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaMoreInfo(Request\Manga\MangaMoreInfoRequest $request): ?string
     {
@@ -678,8 +719,8 @@ class MalClient
     /**
      * @param Request\SeasonList\SeasonListRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getSeasonList(Request\SeasonList\SeasonListRequest $request): array
     {
@@ -696,8 +737,8 @@ class MalClient
     /**
      * @param Request\User\UserHistoryRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getUserHistory(Request\User\UserHistoryRequest $request): array
     {
@@ -713,16 +754,13 @@ class MalClient
 
     /**
      * @param Request\User\UserAnimeListRequest $request
-     *
-     * @return Model\User\AnimeListItem[]
-     * @throws ParserException
+     * @return array
+     * @throws BadResponseException
      */
     public function getUserAnimeList(Request\User\UserAnimeListRequest $request): array
     {
-        $client = new \GuzzleHttp\Client();
-
         try {
-            $response = $client->get($request->getPath());
+            $response = $this->ghoutte->getClient()->get($request->getPath());
             $list = json_decode($response->getBody()->getContents());
 
             $model = [];
@@ -731,21 +769,23 @@ class MalClient
             }
             return $model;
         } catch (\Exception $e) {
-            throw ParserException::fromRequest($request, $e);
+            throw new BadResponseException(
+                $e->getCode().' on '.$request->getPath(),
+                $e->getCode()
+            );
         }
     }
+
 
     /**
      * @param Request\User\UserMangaListRequest $request
      * @return array
-     * @throws ParserException
+     * @throws BadResponseException
      */
     public function getUserMangaList(Request\User\UserMangaListRequest $request): array
     {
-        $client = new \GuzzleHttp\Client();
-
         try {
-            $response = $client->get($request->getPath());
+            $response = $this->ghoutte->getClient()->get($request->getPath());
             $list = json_decode($response->getBody()->getContents());
 
             $model = [];
@@ -754,7 +794,10 @@ class MalClient
             }
             return $model;
         } catch (\Exception $e) {
-            throw ParserException::fromRequest($request, $e);
+            throw new BadResponseException(
+                $e->getCode().' on '.$request->getPath(),
+                $e->getCode()
+            );
         }
     }
 
@@ -762,8 +805,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeRecentlyUpdatedByUsersRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeRecentlyUpdatedByUsers(Request\Anime\AnimeRecentlyUpdatedByUsersRequest $request): array
     {
@@ -780,8 +823,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaRecentlyUpdatedByUsersRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaRecentlyUpdatedByUsers(Request\Manga\MangaRecentlyUpdatedByUsersRequest $request): array
     {
@@ -798,8 +841,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeRecommendationsRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeRecommendations(Request\Anime\AnimeRecommendationsRequest $request): array
     {
@@ -816,8 +859,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaRecommendationsRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaRecommendations(Request\Manga\MangaRecommendationsRequest $request): array
     {
@@ -836,8 +879,8 @@ class MalClient
      * @param Request\Club\UserListRequest $request
      *
      * @return Model\Club\UserProfile[]
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getClubUsers(Request\Club\UserListRequest $request): array
     {
@@ -855,8 +898,8 @@ class MalClient
     /**
      * @param Request\Anime\AnimeReviewsRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getAnimeReviews(Request\Anime\AnimeReviewsRequest $request): array
     {
@@ -873,8 +916,8 @@ class MalClient
     /**
      * @param Request\Manga\MangaReviewsRequest $request
      * @return array
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getMangaReviews(Request\Manga\MangaReviewsRequest $request): array
     {
@@ -892,8 +935,8 @@ class MalClient
     /**
      * @param Request\Club\ClubRequest $request
      * @return Model\Club\Club
+     * @throws BadResponseException
      * @throws ParserException
-     * @throws \HttpResponseException
      */
     public function getClub(Request\Club\ClubRequest $request): Model\Club\Club
     {