+++ src/Parser/User/Profile/UserProfileParser.php
@@ -48,7 +48,7 @@ class UserProfileParser
      */
     public function getUsername(): string
     {
-        return (string)preg_replace('#.*/(\w+)$#', '$1', $this->getProfileUrl());
+        return preg_replace('#.*/(.*)$#', '$1', $this->getProfileUrl());
     }
 
     /**
@@ -85,16 +85,22 @@ class UserProfileParser
         );
     }
 
+
     /**
-     * @return \DateTimeImmutable
-     * @throws \InvalidArgumentException
+     * @return \DateTimeImmutable|null
+     * @throws \Exception
      */
-    public function getLastOnline(): \DateTimeImmutable
+    public function getLastOnline(): ?\DateTimeImmutable
     {
-        return new \DateTimeImmutable(
-            $this->crawler->filterXPath('//span[contains(text(), \'Last Online\')]/following-sibling::span')->text(),
-            new \DateTimeZone('UTC')
-        );
+        try {
+            return new \DateTimeImmutable(
+                $this->crawler->filterXPath('//span[contains(text(), \'Last Online\')]/following-sibling::span')
+                    ->text(),
+                new \DateTimeZone('UTC')
+            );
+        } catch (\Exception $e) {
+            return null;
+        }
     }
 
     /**