+++ src/Request/Search/MangaSearchRequest.php
@@ -2,6 +2,7 @@
 
 namespace Jikan\Request\Search;
 
+use Jikan\Exception\BadResponseException;
 use Jikan\Request\RequestInterface;
 
 /**
@@ -73,6 +74,16 @@ class MangaSearchRequest implements RequestInterface
     private $genreExclude = false;
 
     /**
+     * @var int
+     */
+    private $orderBy;
+
+    /**
+     * @var int
+     */
+    private $sort;
+
+    /**
      * MangaSearchRequest constructor.
      *
      * @param string|null $query
@@ -83,7 +94,13 @@ class MangaSearchRequest implements RequestInterface
         $this->query = $query;
         $this->page = $page;
 
-        $this->query = $this->query ?? "";
+        $this->query = $this->query ?? '';
+
+        $querySize = strlen($this->query);
+
+        if ($querySize > 0 & $querySize < 3) {
+            throw new BadResponseException('Search queries requires at least 3 characters');
+        }
     }
 
     /**
@@ -110,6 +127,8 @@ class MangaSearchRequest implements RequestInterface
                 'em'     => $this->endDate[1],
                 'ey'     => $this->endDate[2],
                 'gx'     => (int)$this->genreExclude,
+                'o'      => $this->orderBy,
+                'w'      => $this->sort
             ]
         );
 
@@ -134,7 +153,7 @@ class MangaSearchRequest implements RequestInterface
     public function setQuery(?string $query = null): self
     {
         $this->query = $query;
-        $this->query = $this->query ?? "";
+        $this->query = $this->query ?? '';
 
         return $this;
     }
@@ -263,6 +282,36 @@ class MangaSearchRequest implements RequestInterface
     {
         $this->genreExclude = $genreExclude;
 
+        return $this;
+    }
+
+    /**
+     * @param string $char
+     * @return MangaSearchRequest
+     */
+    public function setChar(string $char): MangaSearchRequest
+    {
+        $this->char = $char;
+        return $this;
+    }
+
+    /**
+     * @param int $orderBy
+     * @return MangaSearchRequest
+     */
+    public function setOrderBy(int $orderBy): MangaSearchRequest
+    {
+        $this->orderBy = $orderBy;
+        return $this;
+    }
+
+    /**
+     * @param int $sort
+     * @return MangaSearchRequest
+     */
+    public function setSort(int $sort): MangaSearchRequest
+    {
+        $this->sort = $sort;
         return $this;
     }
 }